+++ src/compas_cadwork/datamodel/__init__.py
@@ -1,5 +1,4 @@
 from .element import Element
-from .element import ElementType
 from .element import ElementGroup
 from .element import ElementGroupingType
 from .element import ATTR_INSTRUCTION_ID
@@ -9,7 +8,6 @@ from .dimension import AnchorPoint
 
 __all__ = [
     "Element",
-    "ElementType",
     "ElementGroup",
     "ElementGroupingType",
     "ATTR_INSTRUCTION_ID",