+++ src/compas_cadwork/utilities/events.py
@@ -23,7 +23,7 @@ class ElementDelta:
         new_ids = current_ids - self._known_element_ids
         removed_ids = self._known_element_ids - current_ids
         self._known_element_ids = current_ids
-        return [Element.from_id(id) for id in new_ids], [Element.from_id(id) for id in removed_ids]
+        return [Element(id) for id in new_ids], [Element(id) for id in removed_ids]
 
     def reset(self):
         """Reset the known element ids"""